--- src/main/java/oridungjeol/duckhang/chatting/application/dto/Chat.java
@@ -1,24 +1,0 @@
-package oridungjeol.duckhang.chatting.application.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-public class Chat {
-    private int id;
-
-    private String type;
-
-    private String author_uuid;
-
-    private String content;
-
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
-    private LocalDateTime created_at;
-
-    private int room_id;
-}
